CLIM-1447: Atom14 improve-clarity rename and re-organize large files and move tests
