+++ apps/src/components/map-layers/s2d-variable-values.tsx
@@ -10,15 +10,15 @@ import { useClimateVariable } from '@/hooks/use-climate-variable';
 import { FetchError, fetchS2DLocationData } from '@/services/services';
 
 import { formatValue } from '@/lib/format';
-import { cn, findCeilingIndex, utc } from '@/lib/utils';
+import { cn, utc } from '@/lib/utils';
 import {
 	extractSkillLevelData,
 	generatePeriodRangeLabel,
-	normalizeProbabilitiesBarChartPercent,
+	getProbabilitiesBarChartColour,
 	isFrequencyTypeDecadal,
+	normalizeProbabilitiesBarChartPercent,
 	type LocationS2DData,
 } from '@/lib/s2d';
-import { ColourMap } from '@/types/types';
 import {
 	ForecastDisplay,
 	ForecastDisplays,
@@ -34,7 +34,6 @@ import S2DLocationModalForecastSummary from '@/components/map-layers/s2d-locatio
 
 import ProgressBar from '@/components/ui/progress-bar';
 import type {
-	HexColor,
 	ProgressBarProps,
 } from '@/types/progress-bar';
 
@@ -47,7 +46,7 @@ interface S2DVariableValuesProps {
 	latlng: Pick<L.LatLng, 'lat' | 'lng'>;
 }
 
-interface PopupContentPartProps {
+interface LocationModalContentPartProps {
 	locationData: LocationS2DData | null;
 	dateRangeStart: string | null;
 	frequency: S2DFrequencyType;
@@ -147,111 +146,6 @@ const SKILL_LEVEL_TOOLTIP = [
 	),
 ];
 
-/**
- * Return true if two numbers are in the same thousand.
- */
-const isSameThousand = (valueA: number, valueB: number): boolean =>
-	Math.floor(valueA / 1000) === Math.floor(valueB / 1000);
-
-/**
- * Return true if the given RGB colour code is white.
- *
- * @param colour - The colour code, prefixed with `#`.
- */
-const isWhite = (colour: `#${string}`): boolean =>
-	colour.toUpperCase() === '#FFFFFF' || colour.toUpperCase() === '#FFF';
-
-/**
- * Given a colour map, returns the colour for a specific percentage and outcome.
- *
- * The `colorMap.quantities` defines the boundaries where colours change. For a
- * given quantity (calculated as `1000 * (outcome + 1) + percentage`), the
- * returned colour will be the one associated with the *next* boundary. In other
- * words, the boundaries serve to define the colour for values below them.
- *
- * Special cases:
- * - If the calculated quantity is above the highest boundary for the same
- *   outcome, a default colour is returned.
- * - If the determined colour is white, the next non-white colour in the same
- *   outcome is returned. If no such colour exists, a default colour is
- *   returned.
- * - If the calculated quantity is exactly on a boundary, the colour associated
- *   with the *next* boundary will be returned (in other words, a boundary
- *   quantity is exclusive).
- *   - An exception to the previous case: if the percentage is 100 *and* its
- *     calculated quantity is a boundary, the colour associated with the current
- *     boundary (not the next one) is returned. In other words, 100 is included
- *     in the range that ends at 100. If 100 is not a boundary, the above rules
- *     apply (i.e. a default colour is returned if there is no next boundary in
- *     the same outcome).
- *
- * @param outcome - Index number of the outcome (e.g. 0 for "above", 1 for "below", ...)
- * @param percentage - Percentage for which to get the colour
- * @param colorMap - The colour map containing the colours
- * @returns The colour associated with the given percentage and outcome.
- */
-export const getProbabilityColour = (
-	outcome: number,
-	percentage: number,
-	colorMap: ColourMap
-): HexColor => {
-	const colours = colorMap.colours as HexColor[];
-	const defaultColor: HexColor = '#909090';
-	let percentageForQuery = normalizeProbabilitiesBarChartPercent({ percent: percentage });
-
-	/**
-	 * For values falling exactly on the limit between two colors, we always
-	 * use the color "on the right", except for 100, where we use the color
-	 * on the left.
-	 *
-	 * To achieve this behavior, we add 0.1 to the (rounded) percentage
-	 * (unless if 100) to ensure we always take the color on the right.
-	 */
-	if (percentageForQuery < 100) {
-		percentageForQuery += 0.1;
-	}
-
-	// The "quantity" associated with this percentage and outcome.
-	// For example, an outcome of 0 (e.g. "above") and a `percentageForQuery` of
-	// 23.1 would give 1023.1
-	const queryQuantity = 1000 * (outcome + 1) + percentageForQuery;
-
-	const colourIndex = findCeilingIndex(colorMap.quantities, queryQuantity);
-
-	// If the percentage/outcome is bigger than the highest value
-	if (colourIndex === -1) {
-		return defaultColor;
-	}
-
-	// If the next colour is not in the same outcome
-	if (!isSameThousand(colorMap.quantities[colourIndex], queryQuantity)) {
-		return defaultColor;
-	}
-
-	let colour = colours[colourIndex];
-
-	// If the colour is white, we find the next none-white colour
-	if (isWhite(colour)) {
-		// First, set the colour to a default "grey", in case we don't find
-		// another replacement colour.
-		colour = defaultColor;
-
-		for (let i = colourIndex + 1; i < colours.length; i++) {
-			// Stop if we are now in the next outcome's colours
-			if (!isSameThousand(colorMap.quantities[i], queryQuantity)) {
-				break;
-			}
-
-			const nextColour = colours[i];
-			if (!isWhite(nextColour)) {
-				colour = nextColour;
-				break;
-			}
-		}
-	}
-
-	return colour;
-};
 
 /**
  * Main component for the location popup of a S2D variable. Wrapper containing
@@ -349,7 +243,7 @@ export const S2DVariableValues = (
 	const effectiveLocationData = hasDataLoaded ? locationData : null;
 
 	return (
-		<PopupContent
+		<LocationModalContentPart
 			locationData={effectiveLocationData}
 			dateRangeStart={dateRangeStart}
 			frequency={frequency}
@@ -751,7 +645,7 @@ const ForecastProbabilitiesPart = (
 					),
 					labelTooltipCutoff: '> ' + formatValue(aboveValue, unit, 1, locale),
 					percent: abovePercentage,
-					fillHexCode: getProbabilityColour(
+					fillHexCode: getProbabilitiesBarChartColour(
 						0,
 						abovePercentage,
 						colorMap
@@ -770,7 +664,7 @@ const ForecastProbabilitiesPart = (
 						formatValue(aboveValue, unit, 1, locale)
 					),
 					percent: nearPercentage,
-					fillHexCode: getProbabilityColour(
+					fillHexCode: getProbabilitiesBarChartColour(
 						1,
 						nearPercentage,
 						colorMap
@@ -783,7 +677,7 @@ const ForecastProbabilitiesPart = (
 					),
 					labelTooltipCutoff: '< ' + formatValue(belowValue, unit, 1, locale),
 					percent: belowPercentage,
-					fillHexCode: getProbabilityColour(
+					fillHexCode: getProbabilitiesBarChartColour(
 						2,
 						belowPercentage,
 						colorMap
@@ -804,7 +698,7 @@ const ForecastProbabilitiesPart = (
 					),
 					labelTooltipCutoff: '> ' + formatValue(higherValue, unit, 1, locale),
 					percent: higherPercentage,
-					fillHexCode: getProbabilityColour(
+					fillHexCode: getProbabilitiesBarChartColour(
 						0,
 						higherPercentage,
 						colorMap
@@ -817,7 +711,7 @@ const ForecastProbabilitiesPart = (
 					),
 					labelTooltipCutoff: '< ' + formatValue(lowerValue, unit, 1, locale),
 					percent: lowerPercentage,
-					fillHexCode: getProbabilityColour(
+					fillHexCode: getProbabilitiesBarChartColour(
 						1,
 						lowerPercentage,
 						colorMap
@@ -959,8 +853,8 @@ ForecastProbabilitiesPart.displayName = 'ForecastProbabilitiesPart'; // Explicit
  * @param variableName - The climate variable's name
  * @param unit - The climate variable's unit
  */
-const PopupContentPart = (
-	props: PopupContentPartProps,
+const LocationModalContentPart = (
+	props: LocationModalContentPartProps,
 ): React.ReactElement => {
 	const {
 		locationData,
@@ -1029,4 +923,4 @@ const PopupContentPart = (
 	);
 };
 
-PopupContentPart.displayName = 'PopupContentPart'; // Explicit string literal, or this name would be lost in production.
+LocationModalContentPart.displayName = 'LocationModalContentPart'; // Explicit string literal, or this name would be lost in production.
